Refactor Python SDK: Introduce Morphik SDK, replace DataBridge references, and add exception handling
