+++ examples/cache_augmented_generation.py
@@ -1,12 +1,12 @@
 import os
 from dotenv import load_dotenv
-from databridge import DataBridge
+from morphik import Morphik
 
 # Load environment variables
 load_dotenv()
 
-# Connect to DataBridge
-db = DataBridge(os.getenv("DATABRIDGE_URI"), timeout=10000, is_local=True)
+# Connect to Morphik
+db = Morphik(os.getenv("MORPHIK_URI"), timeout=10000, is_local=True)
 
 # Sample document for demonstration
 long_document = """
