+++ examples/colpali.py
@@ -1,13 +1,13 @@
 import io
-from databridge import DataBridge
+from morphik import Morphik
 import os
 from dotenv import load_dotenv
 from PIL import Image
 
 load_dotenv()
 
-## Connect to the DataBridge instance
-db = DataBridge(os.getenv("DATABRIDGE_URI"), timeout=10000, is_local=True)
+## Connect to the Morphik instance
+db = Morphik(os.getenv("MORPHIK_URI"), timeout=10000, is_local=True)
 
 ## Ingestion Pathway
 db.ingest_file("examples/assets/colpali_example.pdf", use_colpali=True)
@@ -22,6 +22,6 @@ for chunk in chunks:
     else:
         print(chunk.content)
 
-# You can also directly query a VLM as defined in `databridge.toml`
+# You can also directly query a VLM as defined in the configuration
 response = db.query("At what frequency do we achieve the highest Image Rejection Ratio?", use_colpali=True, k=3)
 print(response.completion)
