+++ sdks/python/README.md
@@ -5,14 +5,14 @@ A Python client for Morphik API that enables document ingestion and semantic sea
 ## Installation
 
 ```bash
-pip install databridge-client
+pip install morphik
 ```
 
 ```python
-from databridge import DataBridge
+from morphik import Morphik
 
 # Initialize client
-db = DataBridge("your-api-key")
+db = Morphik("your-api-key")
 
 # Ingest a document
 doc_id = await db.ingest_document(
