--- sdks/python/databridge/__init__.py
@@ -1,15 +1,0 @@
-"""
-DataBridge Python SDK for document ingestion and querying.
-"""
-
-from .sync import DataBridge
-from .async_ import AsyncDataBridge
-from .models import Document
-
-__all__ = [
-    "DataBridge",
-    "AsyncDataBridge",
-    "Document",
-]
-
-__version__ = "0.3.0"
