+++ sdks/python/morphik/async_.py
@@ -1,24 +1,21 @@
-import base64
 from io import BytesIO, IOBase
-import io
-from PIL.Image import Image as PILImage
-from PIL import Image
 import json
 import logging
 from pathlib import Path
 from typing import Dict, Any, List, Optional, Union, BinaryIO
 from urllib.parse import urlparse
 
+import httpx
 import jwt
+from PIL.Image import Image as PILImage
 from pydantic import BaseModel, Field
-import requests
 
 from .models import (
-    Document, 
-    ChunkResult, 
-    DocumentResult, 
-    CompletionResponse, 
-    IngestTextRequest, 
+    Document,
+    ChunkResult,
+    DocumentResult,
+    CompletionResponse,
+    IngestTextRequest,
     ChunkSource,
     Graph,
     # Prompt override models
@@ -38,23 +35,23 @@ logger = logging.getLogger(__name__)
 RuleOrDict = Union[Rule, Dict[str, Any]]
 
 
-class Cache:
-    def __init__(self, db: "DataBridge", name: str):
+class AsyncCache:
+    def __init__(self, db: "AsyncMorphik", name: str):
         self._db = db
         self._name = name
 
-    def update(self) -> bool:
-        response = self._db._request("POST", f"cache/{self._name}/update")
+    async def update(self) -> bool:
+        response = await self._db._request("POST", f"cache/{self._name}/update")
         return response.get("success", False)
 
-    def add_docs(self, docs: List[str]) -> bool:
-        response = self._db._request("POST", f"cache/{self._name}/add_docs", {"docs": docs})
+    async def add_docs(self, docs: List[str]) -> bool:
+        response = await self._db._request("POST", f"cache/{self._name}/add_docs", {"docs": docs})
         return response.get("success", False)
 
-    def query(
+    async def query(
         self, query: str, max_tokens: Optional[int] = None, temperature: Optional[float] = None
     ) -> CompletionResponse:
-        response = self._db._request(
+        response = await self._db._request(
             "POST",
             f"cache/{self._name}/query",
             params={"query": query, "max_tokens": max_tokens, "temperature": temperature},
@@ -77,31 +74,39 @@ class FinalChunkResult(BaseModel):
         arbitrary_types_allowed = True
 
 
-class DataBridge:
+class AsyncMorphik:
     """
-    DataBridge client for document operations.
+    Morphik client for document operations.
 
     Args:
-        uri (str, optional): DataBridge URI in format "databridge://<owner_id>:<token>@<host>".
+        uri (str, optional): Morphik URI in format "morphik://<owner_id>:<token>@<host>".
             If not provided, connects to http://localhost:8000 without authentication.
         timeout (int, optional): Request timeout in seconds. Defaults to 30.
-        is_local (bool, optional): Whether connecting to local development server. Defaults to False.
+        is_local (bool, optional): Whether to connect to a local server. Defaults to False.
 
     Examples:
         ```python
         # Without authentication
-        db = DataBridge()
+        async with AsyncMorphik() as db:
+            doc = await db.ingest_text("Sample content")
 
         # With authentication
-        db = DataBridge("databridge://owner_id:token@api.databridge.ai")
+        async with AsyncMorphik("morphik://owner_id:token@api.morphik.ai") as db:
+            doc = await db.ingest_text("Sample content")
         ```
     """
 
     def __init__(self, uri: Optional[str] = None, timeout: int = 30, is_local: bool = False):
         self._timeout = timeout
-        self._session = requests.Session()
-        if is_local:
-            self._session.verify = False  # Disable SSL for localhost
+        self._client = (
+            httpx.AsyncClient(timeout=timeout)
+            if not is_local
+            else httpx.AsyncClient(
+                timeout=timeout,
+                verify=False,  # Disable SSL for localhost
+                http2=False,  # Force HTTP/1.1
+            )
+        )
         self._is_local = is_local
 
         if uri:
@@ -126,7 +131,7 @@ class DataBridge:
         # Basic token validation
         jwt.decode(self._auth_token, options={"verify_signature": False})
 
-    def _request(
+    async def _request(
         self,
         method: str,
         endpoint: str,
@@ -143,17 +148,16 @@ class DataBridge:
         if files:
             # Multipart form data for files
             request_data = {"files": files, "data": data}
-            # Don't set Content-Type, let requests handle it
+            # Don't set Content-Type, let httpx handle it
         else:
             # JSON for everything else
             headers["Content-Type"] = "application/json"
             request_data = {"json": data}
 
-        response = self._session.request(
+        response = await self._client.request(
             method,
             f"{self._base_url}/{endpoint.lstrip('/')}",
             headers=headers,
-            timeout=self._timeout,
             params=params,
             **request_data,
         )
@@ -166,7 +170,7 @@ class DataBridge:
             return rule.to_dict()
         return rule
 
-    def ingest_text(
+    async def ingest_text(
         self,
         content: str,
         filename: Optional[str] = None,
@@ -175,7 +179,7 @@ class DataBridge:
         use_colpali: bool = True,
     ) -> Document:
         """
-        Ingest a text document into DataBridge.
+        Ingest a text document into Morphik.
 
         Args:
             content: Text content to ingest
@@ -189,7 +193,7 @@ class DataBridge:
 
         Example:
             ```python
-            from databridge.rules import MetadataExtractionRule, NaturalLanguageRule
+            from morphik.rules import MetadataExtractionRule, NaturalLanguageRule
             from pydantic import BaseModel
 
             class DocumentInfo(BaseModel):
@@ -197,7 +201,7 @@ class DataBridge:
                 author: str
                 date: str
 
-            doc = db.ingest_text(
+            doc = await db.ingest_text(
                 "Machine learning is fascinating...",
                 metadata={"category": "tech"},
                 rules=[
@@ -216,72 +220,31 @@ class DataBridge:
             rules=[self._convert_rule(r) for r in (rules or [])],
             use_colpali=use_colpali,
         )
-        response = self._request("POST", "ingest/text", data=request.model_dump())
+        response = await self._request("POST", "ingest/text", data=request.model_dump())
         doc = Document(**response)
         doc._client = self
         return doc
 
-    def ingest_file(
+    async def ingest_file(
         self,
         file: Union[str, bytes, BinaryIO, Path],
-        filename: Optional[str] = None,
+        filename: str,
         metadata: Optional[Dict[str, Any]] = None,
         rules: Optional[List[RuleOrDict]] = None,
         use_colpali: bool = True,
     ) -> Document:
-        """
-        Ingest a file document into DataBridge.
-
-        Args:
-            file: File to ingest (path string, bytes, file object, or Path)
-            filename: Name of the file
-            metadata: Optional metadata dictionary
-            rules: Optional list of rules to apply during ingestion. Can be:
-                  - MetadataExtractionRule: Extract metadata using a schema
-                  - NaturalLanguageRule: Transform content using natural language
-            use_colpali: Whether to use ColPali-style embedding model to ingest the file (slower, but significantly better retrieval accuracy for images)
-
-        Returns:
-            Document: Metadata of the ingested document
-
-        Example:
-            ```python
-            from databridge.rules import MetadataExtractionRule, NaturalLanguageRule
-            from pydantic import BaseModel
-
-            class DocumentInfo(BaseModel):
-                title: str
-                author: str
-                department: str
-
-            doc = db.ingest_file(
-                "document.pdf",
-                filename="document.pdf",
-                metadata={"category": "research"},
-                rules=[
-                    MetadataExtractionRule(schema=DocumentInfo),
-                    NaturalLanguageRule(prompt="Extract key points only")
-                ], # Optional
-                use_colpali=True, # Optional
-            )
-            ```
-        """
+        """Ingest a file document into Morphik."""
         # Handle different file input types
         if isinstance(file, (str, Path)):
             file_path = Path(file)
             if not file_path.exists():
                 raise ValueError(f"File not found: {file}")
-            filename = file_path.name if filename is None else filename
             with open(file_path, "rb") as f:
                 content = f.read()
                 file_obj = BytesIO(content)
         elif isinstance(file, bytes):
-            if filename is None:
-                raise ValueError("filename is required when ingesting bytes")
             file_obj = BytesIO(file)
         else:
-            if filename is None:
-                raise ValueError("filename is required when ingesting file object")
             file_obj = file
 
         try:
@@ -289,14 +252,13 @@ class DataBridge:
             files = {"file": (filename, file_obj)}
 
             # Add metadata and rules
-            form_data = {
+            data = {
                 "metadata": json.dumps(metadata or {}),
                 "rules": json.dumps([self._convert_rule(r) for r in (rules or [])]),
+                "use_colpali": json.dumps(use_colpali),
             }
 
-            response = self._request(
-                "POST", f"ingest/file?use_colpali={str(use_colpali).lower()}", data=form_data, files=files
-            )
+            response = await self._request("POST", "ingest/file", data=data, files=files)
             doc = Document(**response)
             doc._client = self
             return doc
@@ -305,7 +267,7 @@ class DataBridge:
             if isinstance(file, (str, Path)):
                 file_obj.close()
 
-    def ingest_files(
+    async def ingest_files(
         self,
         files: List[Union[str, bytes, BinaryIO, Path]],
         metadata: Optional[Union[Dict[str, Any], List[Dict[str, Any]]]] = None,
@@ -314,7 +276,7 @@ class DataBridge:
         parallel: bool = True,
     ) -> List[Document]:
         """
-        Ingest multiple files into DataBridge.
+        Ingest multiple files into Morphik.
 
         Args:
             files: List of files to ingest (path strings, bytes, file objects, or Paths)
@@ -360,7 +322,7 @@ class DataBridge:
                 "parallel": str(parallel).lower(),
             }
 
-            response = self._request("POST", "ingest/files", data=data, files=file_objects)
+            response = await self._request("POST", "ingest/files", data=data, files=file_objects)
             
             if response.get("errors"):
                 # Log errors but don't raise exception
@@ -377,7 +339,7 @@ class DataBridge:
                 if isinstance(file_obj, (IOBase, BytesIO)) and not file_obj.closed:
                     file_obj.close()
 
-    def ingest_directory(
+    async def ingest_directory(
         self,
         directory: Union[str, Path],
         recursive: bool = False,
@@ -388,7 +350,7 @@ class DataBridge:
         parallel: bool = True,
     ) -> List[Document]:
         """
-        Ingest all files in a directory into DataBridge.
+        Ingest all files in a directory into Morphik.
 
         Args:
             directory: Path to directory containing files to ingest
@@ -422,7 +384,7 @@ class DataBridge:
             return []
 
         # Use ingest_files with collected paths
-        return self.ingest_files(
+        return await self.ingest_files(
             files=files,
             metadata=metadata,
             rules=rules,
@@ -430,7 +392,7 @@ class DataBridge:
             parallel=parallel
         )
 
-    def retrieve_chunks(
+    async def retrieve_chunks(
         self,
         query: str,
         filters: Optional[Dict[str, Any]] = None,
@@ -439,20 +401,20 @@ class DataBridge:
         use_colpali: bool = True,
     ) -> List[FinalChunkResult]:
         """
-        Retrieve relevant chunks.
+        Search for relevant chunks.
 
         Args:
             query: Search query text
             filters: Optional metadata filters
             k: Number of results (default: 4)
             min_score: Minimum similarity threshold (default: 0.0)
-            use_colpali: Whether to use ColPali-style embedding model to retrieve the chunks (only works for documents ingested with `use_colpali=True`)
+            use_colpali: Whether to use ColPali-style embedding model to retrieve chunks (only works for documents ingested with `use_colpali=True`)
         Returns:
-            List[ChunkResult]
+            List[FinalChunkResult]
 
         Example:
             ```python
-            chunks = db.retrieve_chunks(
+            chunks = await db.retrieve_chunks(
                 "What are the key findings?",
                 filters={"department": "research"}
             )
@@ -466,11 +428,10 @@ class DataBridge:
             "use_colpali": use_colpali,
         }
 
-        response = self._request("POST", "retrieve/chunks", request)
+        response = await self._request("POST", "retrieve/chunks", data=request)
         chunks = [ChunkResult(**r) for r in response]
-
+        
         final_chunks = []
-
         for chunk in chunks:
             if chunk.metadata.get("is_image"):
                 try:
@@ -481,12 +442,15 @@ class DataBridge:
                         content = content.split(",", 1)[1]
 
                     # Now decode the base64 string
+                    import base64
+                    import io
+                    from PIL import Image
                     image_bytes = base64.b64decode(content)
                     content = Image.open(io.BytesIO(image_bytes))
                 except Exception as e:
                     print(f"Error processing image: {str(e)}")
                     # Fall back to using the content as text
-                    print(chunk.content)
+                    content = chunk.content
             else:
                 content = chunk.content
 
@@ -502,10 +466,10 @@ class DataBridge:
                     download_url=chunk.download_url,
                 )
             )
-
+            
         return final_chunks
 
-    def retrieve_docs(
+    async def retrieve_docs(
         self,
         query: str,
         filters: Optional[Dict[str, Any]] = None,
@@ -521,13 +485,13 @@ class DataBridge:
             filters: Optional metadata filters
             k: Number of results (default: 4)
             min_score: Minimum similarity threshold (default: 0.0)
-            use_colpali: Whether to use ColPali-style embedding model to retrieve the documents (only works for documents ingested with `use_colpali=True`)
+            use_colpali: Whether to use ColPali-style embedding model to retrieve documents (only works for documents ingested with `use_colpali=True`)
         Returns:
             List[DocumentResult]
 
         Example:
             ```python
-            docs = db.retrieve_docs(
+            docs = await db.retrieve_docs(
                 "machine learning",
                 k=5
             )
@@ -541,10 +505,10 @@ class DataBridge:
             "use_colpali": use_colpali,
         }
 
-        response = self._request("POST", "retrieve/docs", request)
+        response = await self._request("POST", "retrieve/docs", data=request)
         return [DocumentResult(**r) for r in response]
 
-    def query(
+    async def query(
         self,
         query: str,
         filters: Optional[Dict[str, Any]] = None,
@@ -580,14 +544,14 @@ class DataBridge:
         Example:
             ```python
             # Standard query
-            response = db.query(
+            response = await db.query(
                 "What are the key findings about customer satisfaction?",
                 filters={"department": "research"},
                 temperature=0.7
             )
             
             # Knowledge graph enhanced query
-            response = db.query(
+            response = await db.query(
                 "How does product X relate to customer segment Y?",
                 graph_name="market_graph",
                 hop_depth=2,
@@ -595,8 +559,8 @@ class DataBridge:
             )
             
             # With prompt customization
-            from databridge.models import QueryPromptOverride, QueryPromptOverrides
-            response = db.query(
+            from morphik.models import QueryPromptOverride, QueryPromptOverrides
+            response = await db.query(
                 "What are the key findings?",
                 prompt_overrides=QueryPromptOverrides(
                     query=QueryPromptOverride(
@@ -606,7 +570,7 @@ class DataBridge:
             )
             
             # Or using a dictionary
-            response = db.query(
+            response = await db.query(
                 "What are the key findings?",
                 prompt_overrides={
                     "query": {
@@ -641,10 +605,10 @@ class DataBridge:
             "prompt_overrides": prompt_overrides,
         }
 
-        response = self._request("POST", "query", request)
+        response = await self._request("POST", "query", data=request)
         return CompletionResponse(**response)
 
-    def list_documents(
+    async def list_documents(
         self, skip: int = 0, limit: int = 100, filters: Optional[Dict[str, Any]] = None
     ) -> List[Document]:
         """
@@ -661,20 +625,22 @@ class DataBridge:
         Example:
             ```python
             # Get first page
-            docs = db.list_documents(limit=10)
+            docs = await db.list_documents(limit=10)
 
             # Get next page
-            next_page = db.list_documents(skip=10, limit=10, filters={"department": "research"})
+            next_page = await db.list_documents(skip=10, limit=10, filters={"department": "research"})
             ```
         """
         # Use query params for pagination and POST body for filters
-        response = self._request("POST", f"documents?skip={skip}&limit={limit}", data=filters or {})
+        response = await self._request(
+            "POST", f"documents?skip={skip}&limit={limit}", data=filters or {}
+        )
         docs = [Document(**doc) for doc in response]
         for doc in docs:
             doc._client = self
         return docs
 
-    def get_document(self, document_id: str) -> Document:
+    async def get_document(self, document_id: str) -> Document:
         """
         Get document metadata by ID.
 
@@ -686,16 +652,16 @@ class DataBridge:
 
         Example:
             ```python
-            doc = db.get_document("doc_123")
+            doc = await db.get_document("doc_123")
             print(f"Title: {doc.metadata.get('title')}")
             ```
         """
-        response = self._request("GET", f"documents/{document_id}")
+        response = await self._request("GET", f"documents/{document_id}")
         doc = Document(**response)
         doc._client = self
         return doc
         
-    def get_document_by_filename(self, filename: str) -> Document:
+    async def get_document_by_filename(self, filename: str) -> Document:
         """
         Get document metadata by filename.
         If multiple documents have the same filename, returns the most recently updated one.
@@ -708,16 +674,16 @@ class DataBridge:
 
         Example:
             ```python
-            doc = db.get_document_by_filename("report.pdf")
+            doc = await db.get_document_by_filename("report.pdf")
             print(f"Document ID: {doc.external_id}")
             ```
         """
-        response = self._request("GET", f"documents/filename/{filename}")
+        response = await self._request("GET", f"documents/filename/{filename}")
         doc = Document(**response)
         doc._client = self
         return doc
         
-    def update_document_with_text(
+    async def update_document_with_text(
         self,
         document_id: str,
         content: str,
@@ -729,7 +695,7 @@ class DataBridge:
     ) -> Document:
         """
         Update a document with new text content using the specified strategy.
-
+        
         Args:
             document_id: ID of the document to update
             content: The new content to add
@@ -738,14 +704,14 @@ class DataBridge:
             rules: Optional list of rules to apply to the content
             update_strategy: Strategy for updating the document (currently only 'add' is supported)
             use_colpali: Whether to use multi-vector embedding
-
+            
         Returns:
             Document: Updated document metadata
-
+            
         Example:
             ```python
             # Add new content to an existing document
-            updated_doc = db.update_document_with_text(
+            updated_doc = await db.update_document_with_text(
                 document_id="doc_123",
                 content="This is additional content that will be appended to the document.",
                 filename="updated_document.txt",
@@ -768,7 +734,7 @@ class DataBridge:
         if update_strategy != "add":
             params["update_strategy"] = update_strategy
             
-        response = self._request(
+        response = await self._request(
             "POST", 
             f"documents/{document_id}/update_text", 
             data=request.model_dump(),
@@ -778,8 +744,8 @@ class DataBridge:
         doc = Document(**response)
         doc._client = self
         return doc
-
-    def update_document_with_file(
+        
+    async def update_document_with_file(
         self,
         document_id: str,
         file: Union[str, bytes, BinaryIO, Path],
@@ -791,7 +757,7 @@ class DataBridge:
     ) -> Document:
         """
         Update a document with content from a file using the specified strategy.
-
+        
         Args:
             document_id: ID of the document to update
             file: File to add (path string, bytes, file object, or Path)
@@ -800,14 +766,14 @@ class DataBridge:
             rules: Optional list of rules to apply to the content
             update_strategy: Strategy for updating the document (currently only 'add' is supported)
             use_colpali: Whether to use multi-vector embedding
-
+            
         Returns:
             Document: Updated document metadata
-
+            
         Example:
             ```python
             # Add content from a file to an existing document
-            updated_doc = db.update_document_with_file(
+            updated_doc = await db.update_document_with_file(
                 document_id="doc_123",
                 file="path/to/update.pdf",
                 metadata={"status": "updated"},
@@ -849,7 +815,7 @@ class DataBridge:
                 form_data["use_colpali"] = str(use_colpali).lower()
                 
             # Use the dedicated file update endpoint
-            response = self._request(
+            response = await self._request(
                 "POST", f"documents/{document_id}/update_file", data=form_data, files=files
             )
             
@@ -861,7 +827,7 @@ class DataBridge:
             if isinstance(file, (str, Path)):
                 file_obj.close()
     
-    def update_document_metadata(
+    async def update_document_metadata(
         self,
         document_id: str,
         metadata: Dict[str, Any],
@@ -879,7 +845,7 @@ class DataBridge:
         Example:
             ```python
             # Update just the metadata of a document
-            updated_doc = db.update_document_metadata(
+            updated_doc = await db.update_document_metadata(
                 document_id="doc_123",
                 metadata={"status": "reviewed", "reviewer": "Jane Smith"}
             )
@@ -887,12 +853,12 @@ class DataBridge:
             ```
         """
         # Use the dedicated metadata update endpoint
-        response = self._request("POST", f"documents/{document_id}/update_metadata", data=metadata)
+        response = await self._request("POST", f"documents/{document_id}/update_metadata", data=metadata)
         doc = Document(**response)
         doc._client = self
         return doc
         
-    def update_document_by_filename_with_text(
+    async def update_document_by_filename_with_text(
         self,
         filename: str,
         content: str,
@@ -920,7 +886,7 @@ class DataBridge:
         Example:
             ```python
             # Add new content to an existing document identified by filename
-            updated_doc = db.update_document_by_filename_with_text(
+            updated_doc = await db.update_document_by_filename_with_text(
                 filename="report.pdf",
                 content="This is additional content that will be appended to the document.",
                 new_filename="updated_report.pdf",
@@ -931,10 +897,10 @@ class DataBridge:
             ```
         """
         # First get the document by filename to obtain its ID
-        doc = self.get_document_by_filename(filename)
+        doc = await self.get_document_by_filename(filename)
         
         # Then use the regular update_document_with_text endpoint with the document ID
-        return self.update_document_with_text(
+        return await self.update_document_with_text(
             document_id=doc.external_id,
             content=content,
             filename=new_filename,
@@ -944,7 +910,7 @@ class DataBridge:
             use_colpali=use_colpali
         )
         
-    def update_document_by_filename_with_file(
+    async def update_document_by_filename_with_file(
         self,
         filename: str,
         file: Union[str, bytes, BinaryIO, Path],
@@ -972,7 +938,7 @@ class DataBridge:
         Example:
             ```python
             # Add content from a file to an existing document identified by filename
-            updated_doc = db.update_document_by_filename_with_file(
+            updated_doc = await db.update_document_by_filename_with_file(
                 filename="report.pdf",
                 file="path/to/update.pdf",
                 metadata={"status": "updated"},
@@ -982,10 +948,10 @@ class DataBridge:
             ```
         """
         # First get the document by filename to obtain its ID
-        doc = self.get_document_by_filename(filename)
+        doc = await self.get_document_by_filename(filename)
         
         # Then use the regular update_document_with_file endpoint with the document ID
-        return self.update_document_with_file(
+        return await self.update_document_with_file(
             document_id=doc.external_id,
             file=file,
             filename=new_filename,
@@ -995,7 +961,7 @@ class DataBridge:
             use_colpali=use_colpali
         )
                 
-    def update_document_by_filename_metadata(
+    async def update_document_by_filename_metadata(
         self,
         filename: str,
         metadata: Dict[str, Any],
@@ -1015,7 +981,7 @@ class DataBridge:
         Example:
             ```python
             # Update just the metadata of a document identified by filename
-            updated_doc = db.update_document_by_filename_metadata(
+            updated_doc = await db.update_document_by_filename_metadata(
                 filename="report.pdf",
                 metadata={"status": "reviewed", "reviewer": "Jane Smith"},
                 new_filename="reviewed_report.pdf"  # Optional: rename the file
@@ -1024,10 +990,10 @@ class DataBridge:
             ```
         """
         # First get the document by filename to obtain its ID
-        doc = self.get_document_by_filename(filename)
+        doc = await self.get_document_by_filename(filename)
         
         # Update the metadata
-        result = self.update_document_metadata(
+        result = await self.update_document_metadata(
             document_id=doc.external_id,
             metadata=metadata,
         )
@@ -1038,7 +1004,7 @@ class DataBridge:
             combined_metadata = result.metadata.copy()
             
             # Update the document again with filename change and the same metadata
-            response = self._request(
+            response = await self._request(
                 "POST", 
                 f"documents/{doc.external_id}/update_text", 
                 data={
@@ -1052,8 +1018,8 @@ class DataBridge:
             result._client = self
             
         return result
-        
-    def batch_get_documents(self, document_ids: List[str]) -> List[Document]:
+    
+    async def batch_get_documents(self, document_ids: List[str]) -> List[Document]:
         """
         Retrieve multiple documents by their IDs in a single batch operation.
         
@@ -1065,18 +1031,18 @@ class DataBridge:
             
         Example:
             ```python
-            docs = db.batch_get_documents(["doc_123", "doc_456", "doc_789"])
+            docs = await db.batch_get_documents(["doc_123", "doc_456", "doc_789"])
             for doc in docs:
                 print(f"Document {doc.external_id}: {doc.metadata.get('title')}")
             ```
         """
-        response = self._request("POST", "batch/documents", data=document_ids)
+        response = await self._request("POST", "batch/documents", data=document_ids)
         docs = [Document(**doc) for doc in response]
         for doc in docs:
             doc._client = self
         return docs
         
-    def batch_get_chunks(self, sources: List[Union[ChunkSource, Dict[str, Any]]]) -> List[FinalChunkResult]:
+    async def batch_get_chunks(self, sources: List[Union[ChunkSource, Dict[str, Any]]]) -> List[FinalChunkResult]:
         """
         Retrieve specific chunks by their document ID and chunk number in a single batch operation.
         
@@ -1095,13 +1061,13 @@ class DataBridge:
             ]
             
             # Or using ChunkSource objects
-            from databridge.models import ChunkSource
+            from morphik.models import ChunkSource
             sources = [
                 ChunkSource(document_id="doc_123", chunk_number=0),
                 ChunkSource(document_id="doc_456", chunk_number=2)
             ]
             
-            chunks = db.batch_get_chunks(sources)
+            chunks = await db.batch_get_chunks(sources)
             for chunk in chunks:
                 print(f"Chunk from {chunk.document_id}, number {chunk.chunk_number}: {chunk.content[:50]}...")
             ```
@@ -1114,7 +1080,7 @@ class DataBridge:
             else:
                 source_dicts.append(source.model_dump())
                 
-        response = self._request("POST", "batch/chunks", data=source_dicts)
+        response = await self._request("POST", "batch/chunks", data=source_dicts)
         chunks = [ChunkResult(**r) for r in response]
         
         final_chunks = []
@@ -1128,6 +1094,9 @@ class DataBridge:
                         content = content.split(",", 1)[1]
 
                     # Now decode the base64 string
+                    import base64
+                    import io
+                    from PIL import Image
                     image_bytes = base64.b64decode(content)
                     content = Image.open(io.BytesIO(image_bytes))
                 except Exception as e:
@@ -1152,7 +1121,7 @@ class DataBridge:
             
         return final_chunks
 
-    def create_cache(
+    async def create_cache(
         self,
         name: str,
         model: str,
@@ -1178,7 +1147,7 @@ class DataBridge:
             # This will include both:
             # 1. Any documents with category="programming"
             # 2. The specific documents "doc1" and "doc2" (regardless of their category)
-            cache = db.create_cache(
+            cache = await db.create_cache(
                 name="programming_cache",
                 model="llama2",
                 gguf_file="llama-2-7b-chat.Q4_K_M.gguf",
@@ -1193,10 +1162,10 @@ class DataBridge:
         # Build request body for filters and docs
         request = {"filters": filters, "docs": docs}
 
-        response = self._request("POST", "cache/create", request, params=params)
+        response = await self._request("POST", "cache/create", request, params=params)
         return response
 
-    def get_cache(self, name: str) -> Cache:
+    async def get_cache(self, name: str) -> AsyncCache:
         """
         Get a cache by name.
 
@@ -1208,15 +1177,15 @@ class DataBridge:
 
         Example:
             ```python
-            cache = db.get_cache("programming_cache")
+            cache = await db.get_cache("programming_cache")
             ```
         """
-        response = self._request("GET", f"cache/{name}")
+        response = await self._request("GET", f"cache/{name}")
         if response.get("exists", False):
-            return Cache(self, name)
+            return AsyncCache(self, name)
         raise ValueError(f"Cache '{name}' not found")
 
-    def create_graph(
+    async def create_graph(
         self,
         name: str,
         filters: Optional[Dict[str, Any]] = None,
@@ -1242,20 +1211,20 @@ class DataBridge:
         Example:
             ```python
             # Create a graph from documents with category="research"
-            graph = db.create_graph(
+            graph = await db.create_graph(
                 name="research_graph",
                 filters={"category": "research"}
             )
 
             # Create a graph from specific documents
-            graph = db.create_graph(
+            graph = await db.create_graph(
                 name="custom_graph",
                 documents=["doc1", "doc2", "doc3"]
             )
             
             # With custom entity extraction examples
-            from databridge.models import EntityExtractionPromptOverride, EntityExtractionExample, GraphPromptOverrides
-            graph = db.create_graph(
+            from morphik.models import EntityExtractionPromptOverride, EntityExtractionExample, GraphPromptOverrides
+            graph = await db.create_graph(
                 name="medical_graph", 
                 filters={"category": "medical"},
                 prompt_overrides=GraphPromptOverrides(
@@ -1280,10 +1249,10 @@ class DataBridge:
             "prompt_overrides": prompt_overrides,
         }
 
-        response = self._request("POST", "graph/create", request)
+        response = await self._request("POST", "graph/create", request)
         return Graph(**response)
-        
-    def get_graph(self, name: str) -> Graph:
+
+    async def get_graph(self, name: str) -> Graph:
         """
         Get a graph by name.
 
@@ -1296,14 +1265,14 @@ class DataBridge:
         Example:
             ```python
             # Get a graph by name
-            graph = db.get_graph("finance_graph")
+            graph = await db.get_graph("finance_graph")
             print(f"Graph has {len(graph.entities)} entities and {len(graph.relationships)} relationships")
             ```
         """
-        response = self._request("GET", f"graph/{name}")
+        response = await self._request("GET", f"graph/{name}")
         return Graph(**response)
 
-    def list_graphs(self) -> List[Graph]:
+    async def list_graphs(self) -> List[Graph]:
         """
         List all graphs the user has access to.
 
@@ -1313,15 +1282,15 @@ class DataBridge:
         Example:
             ```python
             # List all accessible graphs
-            graphs = db.list_graphs()
+            graphs = await db.list_graphs()
             for graph in graphs:
                 print(f"Graph: {graph.name}, Entities: {len(graph.entities)}")
             ```
         """
-        response = self._request("GET", "graphs")
+        response = await self._request("GET", "graphs")
         return [Graph(**graph) for graph in response]
-        
-    def update_graph(
+
+    async def update_graph(
         self,
         name: str,
         additional_filters: Optional[Dict[str, Any]] = None,
@@ -1330,33 +1299,33 @@ class DataBridge:
     ) -> Graph:
         """
         Update an existing graph with new documents.
-        
+
         This method processes additional documents matching the original or new filters,
         extracts entities and relationships, and updates the graph with new information.
-        
+
         Args:
             name: Name of the graph to update
             additional_filters: Optional additional metadata filters to determine which new documents to include
             additional_documents: Optional list of additional document IDs to include
             prompt_overrides: Optional customizations for entity extraction and resolution prompts
                 Either a GraphPromptOverrides object or a dictionary with the same structure
-            
+
         Returns:
             Graph: The updated graph
-            
+
         Example:
             ```python
             # Update a graph with new documents
-            updated_graph = db.update_graph(
+            updated_graph = await db.update_graph(
                 name="research_graph",
                 additional_filters={"category": "new_research"},
                 additional_documents=["doc4", "doc5"]
             )
             print(f"Graph now has {len(updated_graph.entities)} entities")
-            
+
             # With entity resolution examples
-            from databridge.models import EntityResolutionPromptOverride, EntityResolutionExample, GraphPromptOverrides
-            updated_graph = db.update_graph(
+            from morphik.models import EntityResolutionPromptOverride, EntityResolutionExample, GraphPromptOverrides
+            updated_graph = await db.update_graph(
                 name="research_graph",
                 additional_documents=["doc4"],
                 prompt_overrides=GraphPromptOverrides(
@@ -1382,10 +1351,10 @@ class DataBridge:
             "prompt_overrides": prompt_overrides,
         }
 
-        response = self._request("POST", f"graph/{name}/update", request)
+        response = await self._request("POST", f"graph/{name}/update", request)
         return Graph(**response)
         
-    def delete_document(self, document_id: str) -> Dict[str, str]:
+    async def delete_document(self, document_id: str) -> Dict[str, str]:
         """
         Delete a document and all its associated data.
         
@@ -1403,14 +1372,14 @@ class DataBridge:
         Example:
             ```python
             # Delete a document
-            result = db.delete_document("doc_123")
+            result = await db.delete_document("doc_123")
             print(result["message"])  # Document doc_123 deleted successfully
             ```
         """
-        response = self._request("DELETE", f"documents/{document_id}")
+        response = await self._request("DELETE", f"documents/{document_id}")
         return response
         
-    def delete_document_by_filename(self, filename: str) -> Dict[str, str]:
+    async def delete_document_by_filename(self, filename: str) -> Dict[str, str]:
         """
         Delete a document by its filename.
         
@@ -1426,22 +1395,22 @@ class DataBridge:
         Example:
             ```python
             # Delete a document by filename
-            result = db.delete_document_by_filename("report.pdf")
+            result = await db.delete_document_by_filename("report.pdf")
             print(result["message"])
             ```
         """
         # First get the document by filename to obtain its ID
-        doc = self.get_document_by_filename(filename)
+        doc = await self.get_document_by_filename(filename)
         
         # Then delete the document by ID
-        return self.delete_document(doc.external_id)
+        return await self.delete_document(doc.external_id)
 
-    def close(self):
-        """Close the HTTP session"""
-        self._session.close()
+    async def close(self):
+        """Close the HTTP client"""
+        await self._client.aclose()
 
-    def __enter__(self):
+    async def __aenter__(self):
         return self
 
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        self.close()
+    async def __aexit__(self, exc_type, exc_val, exc_tb):
+        await self.close()
