+++ sdks/python/morphik/exceptions.py
@@ -1,0 +1,16 @@
+class MorphikError(Exception):
+    """Base exception for Morphik SDK"""
+
+    pass
+
+
+class AuthenticationError(MorphikError):
+    """Authentication related errors"""
+
+    pass
+
+
+class ConnectionError(MorphikError):
+    """Connection related errors"""
+
+    pass
